requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-socketio==5.3.6
eventlet==0.35.1
pyyaml==6.0.1
requests==2.31.0
scikit-learn==1.4.0
numpy==1.26.3
gunicorn==21.2.0
python-dotenv==1.0.0
